pyproject.toml
[tool.poetry]
name = "ej-platform"
version = "3.4.0"
description = "A Django-powered participation platform"
authors = ["Fábio Macêdo Mendes <fabiomacedomendes@gmail.com>, Pencillabs <contato@pencillabs.tec.br>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = ">= 3.9.19, <3.10"
django = "4.1.13"
django-constance = {version = "^2.3", extras = ["database"]}
django-allauth = "0.59.0"
django-taggit = "3.0.0"
django-autoslug = "^1.9"
django-environ = "^0.4.5"
django-bulk-update = "^2.2"
django-manager-utils = "^3.0.1"
django-polymorphic = "^2.0"
django-pandas = "^0.5.1"
django-filter = "^1.1"
django-model-utils = {git = "https://github.com/jazzband/django-model-utils.git", rev = "8e19699b0f8edbb93863de5c72256017080f2af0"}
coreapi = "^2.3"
hyperpython = "^1.1.1"
rules = "<2.0"
sidekick = "==0.5.2"
scipy = "^1.2"
pandas = "^1.2"
scikit-learn = "^1.0"
matplotlib = "^3.0"
invoke = "2.0.0"
wordcloud = "^1.5"
stop-words = "^2018.7"
djangorestframework = "3.14.0"
django-cors-headers = "^3.3.0"
babel = "^2.9.0"
sphinx = "5.0.2"
psycopg2 = "==2.8.6"
sphinx-rtd-theme = "<2.0.0"
dj-rest-auth = ">=2.1.11"
toml = "^0.10.0"
gunicorn = "^19.9"
jinja2 = "2.11.3"
django-debug-toolbar = "3.2.0"
setuptools = "61.2.0"
django-anymail = {extras = ["mailgun"], version = "^10.0"}
django-ckeditor = "6.3.2"
colorama = "^0.4.6"
django-boogie = {git = "https://gitlab.com/pencillabs/ej/django-boogie.git", rev = "fixes-python39"}
importlib-metadata = "^7.1.0"
argon2-cffi = "^23.1.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
numpy = "^1.26.4"
libsass = "^0.23.0"

[tool.poetry.dev-dependencies]
bs4 = "^0.0.1"
mock = "^2.0"
model_mommy = "^1.6"
pytest = "==7.0.0"
pytest-django = "^3.4"
black = {version = "^22.3.0", allow-prereleases = true}
ruff = "0.3.7"
ipython = "^7.2"
watchdog = "^0.9.0"
pre-commit = "^1.17"
notebook = "^6.0"
django-debug-toolbar = "3.2.0"
faker = "^24.4.0"

[tool.black]
line-length = 90
target_version = ['py38']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.ej.conf]
themes=["default"]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ej.settings"
django_find_project = true
pythonpath = "src"
testpaths = "src"
norecursedirs = [".tox", "lib"]
addopts = "--nomigrations"

[tool.ruff.lint.per-file-ignores]
"etc/tasklib/base.py" = ["E731"]
"etc/tasklib/build_tasks.py" = ["E731"]
"src/ej/all.py" = ["E731"]
"src/ej/forms.py" = ["E731"]
"src/ej/roles/tags.py" = ["E731", "F811"]
"src/ej/settings/options.py" = ["E731"]
"src/ej_clusters/math/kmeans.py" = ["E731"]
"src/ej_clusters/models/__init__.py" = ["E731", "F401"]
"src/ej_clusters/models/querysets.py" = ["E731"]
"src/ej_clusters/models/stereotype.py" = ["E731"]
"src/ej_conversations/admin.py" = ["E731"]
"src/ej_conversations/management/commands/_examples.py" = ["E731"]
"src/ej_conversations/models/comment_queryset.py" = ["E731"]
"src/ej_conversations/models/conversation_queryset.py" = ["E731"]
"src/ej_conversations/models/vote.py" = ["E731"]
"src/ej_conversations/models/vote_queryset.py" = ["E731"]
"src/ej_profiles/admin.py" = ["E731"]
"src/ej_profiles/enums.py" = ["E731"]
"src/ej_profiles/management/commands/fixprofiles.py" = ["E731"]
"src/ej_clusters/models/cluster_queryset.py" = ["E731"]
"src/ej_conversations/roles/__init__.py" = ["F401"]
"src/ej_conversations/models/__init__.py" = ["F401"]
"src/ej_conversations/api.py" = ["F401", "F811"]
"src/ej_conversations/__init__.py" = ["F401"]
"src/ej_clusters/tests/conftest.py" = ["F401", "F403"]
"src/ej/components/__init__.py" = ["F401"]
"src/ej_clusters/math/__init__.py" = ["F401"]
"src/ej_users/apps.py" = ["F401"]
"src/ej/testing/__init__.py" = ["F401"]
"src/ej_boards/__init__.py" = ["F401"]
"src/ej/roles/__init__.py" = ["F401", "F403"]
"etc/tasklib/tasks.py" = ["F403"]
"src/ej/tests/__init__.py" = ["F403"]
"src/ej_dataviz/utils.py" = ["F403", "F405"]
"src/ej_dataviz/views_dataviz.py" = ["F403", "F405", "F601", "F811"]
"src/ej_tools/tests/test_routes.py" = ["F403", "F405"]
"tasks.py" = ["F403"]
"src/ej_tools/tests/conftest.py" = ["F403"]
"src/ej_tools/mailing.py" = ["F523"]
"etc/scripts/compilemessages.py" = ["E741"]
"src/ej_dataviz/tests/test_routes.py" = ["F811"]
"src/ej_conversations/views_comments.py" = ["F811"]
"src/ej_conversations/tests/test_api.py" = ["E721"]
"src/ej_boards/migrations/0004_set_board_on_conversations.py" = ["F841"]
"src/ej/tests/pytest/test_urls.py" = ["F821"]
